runtime: move interaction.ts into discord.ts, docs and cleanups
