+++ components/runtime/src/ts/index.ts
@@ -8,6 +8,5 @@ export * from './script';
 export * from './storage';
 export * from './httpclient';
 export * from './scheduled_tasks';
-export * from './interaction';
 export * as Discord from './discord';
 export * as Unstable from './unstable/index';
